move files to proper folder
